some changes for markup
